refactor(orchestrator): split 3324-line module into a package
orchestrator.py mixed five concerns the test suite and the running
state machine actually keep separate: constants + regex patterns,
markdown section parsers, runtime data structures, pure module-level
helpers (time, sort, hook env, dispatch eligibility predicates), and
the long Orchestrator class itself. Splitting along those seams:

  orchestrator/
    constants.py    77 — AUTO_TRIAGE_*, retry timing, STALL_FORCE_EJECT_GRACE_S,
                         _REWIND_TRANSITIONS, all markdown regexes
    parsing.py     136 — _section_body, _parse_touched_files, _parse_findings_rows
    entries.py     125 — RunningEntry, RetryEntry, _CodexTotals, _IssueDebug
    helpers.py     221 — _is_rewind_transition, _branch_hook_env,
                         _branch_already_merged_into_target, _requested_agent_kind,
                         _is_auto_triage_todo_candidate, _config_for_issue_agent,
                         _utc_iso_z, _to_iso, _from_monotonic_to_iso,
                         _max_turns_exhausted_target_state, _notify_state_transition,
                         _task_debug, _sort_for_dispatch_fifo
    core.py       2873 — the Orchestrator class itself (one cohesive unit:
                         tick loop, dispatch, worker coroutine, codex events,
                         reconcile, retry scheduler, startup cleanup). Single
                         class with shared race-prone state — splitting across
                         mixins would risk the memory-tracked slot-leak,
                         done-callback-identity, and two-stage stall invariants.
    __init__.py     88 — re-exports + binds the three monkeypatch targets
                         (build_backend, commit_workspace_on_done,
                         auto_merge_on_done_best_effort) at the package
                         attribute level

Monkeypatch-safe: core.py reads those three names via
`_pkg.<name>` where `_pkg = sys.modules[__package__]`, so test
patches via `monkeypatch.setattr("symphony.orchestrator.X", stub)`
intercept the live call sites just like they did on the flat module.

Public surface unchanged: `Orchestrator`, `RunningEntry`, `_IssueDebug`,
`_sort_for_dispatch_fifo`, `_is_rewind_transition`,
`STALL_FORCE_EJECT_GRACE_S`, `PAUSED_RETRY_HOLD_MS` all still
importable from `symphony.orchestrator`. Race invariants preserved
bit-for-bit (done-callback task-identity check, two-stage stall
force-eject, retry-pending slot accounting, last_progress_timestamp
predicate filtering on type==assistant). pytest 565 passed, 6 skipped
— identical to baseline.
